requirements.txt
pytest
pytest-asyncio
playwright
dotenv # For loading ENVATO_USERNAME and ENVATO_PASSWORD
# Added for Pydantic models, used by Envato client's search params
pydantic
# Added for slugifying keywords in URL paths for Envato client
python-slugify
openai-whisper
